requirements.txt
# Python dependencies for MVP Video System
elevenlabs==0.2.28
Pillow==10.4.0
requests==2.32.3
python-pptx==0.6.23
reportlab==4.2.2
opencv-python==4.10.0.84
numpy==1.26.4
scipy==1.14.0
moviepy==1.0.3
librosa==0.10.2
soundfile==0.12.1
whisper-openai==1.1.10
transformers==4.44.0
torch==2.4.0
torchvision==0.19.0
matplotlib==3.9.2
seaborn==0.13.2
pandas==2.2.2
scikit-learn==1.5.1
boto3==1.34.131
redis==5.0.8
celery==5.3.4
pydub==0.25.1
pytesseract==0.3.10
pyyaml==6.0.2
python-dotenv==1.0.1
fastapi==0.112.0
uvicorn==0.30.1
pydantic==2.8.2
aiofiles==24.1.0
httpx==0.27.0
websockets==12.0
sqlalchemy==2.0.31
alembic==1.13.1
psycopg2-binary==2.9.9
mysql-connector-python==9.0.0
pymongo==4.8.0
motor==3.5.1
jinja2==3.1.4
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.2.0
pytest==8.3.2
pytest-asyncio==0.23.8
black==24.8.0
flake8==7.1.0
mypy==1.11.1
sphinx==7.4.7